+++ examples/actions/TabularInputAction.php
@@ -6,10 +6,10 @@
  * @license https://github.com/unclead/yii2-multiple-input/blob/master/LICENSE.md
  */
 
-namespace yii\multipleinput\examples\actions;
+namespace unclead\multipleinput\examples\actions;
 
 use Yii;
-use yii\multipleinput\examples\models\Item;
+use unclead\multipleinput\examples\models\Item;
 use yii\base\Action;
 use yii\base\Model;
 use yii\bootstrap\ActiveForm;
@@ -17,7 +17,7 @@ use yii\web\Response;
 
 /**
  * Class TabularInputAction
- * @package yii\multipleinput\examples\actions
+ * @package unclead\multipleinput\examples\actions
  */
 class TabularInputAction extends Action
 {