+++ examples/views/multiple-input.php
@@ -1,10 +1,10 @@
 <?php
 
 use yii\bootstrap\ActiveForm;
-use yii\multipleinput\MultipleInput;
-use yii\multipleinput\examples\models\ExampleModel;
+use unclead\multipleinput\MultipleInput;
+use unclead\multipleinput\examples\models\ExampleModel;
 use yii\helpers\Html;
-use yii\multipleinput\MultipleInputColumn;
+use unclead\multipleinput\MultipleInputColumn;
 
 // Note: You have to install https://github.com/kartik-v/yii2-widget-datepicker for correct work an example
 use kartik\date\DatePicker;