+++ src/MultipleInput.php
@@ -6,15 +6,15 @@
  * @license https://github.com/unclead/yii2-multiple-input/blob/master/LICENSE.md
  */
 
-namespace yii\multipleinput;
+namespace unclead\multipleinput;
 
 use Yii;
 use yii\base\Model;
 use yii\helpers\ArrayHelper;
 use yii\widgets\InputWidget;
 use yii\db\ActiveRecordInterface;
-use yii\multipleinput\renderers\TableRenderer;
-use yii\multipleinput\renderers\RendererInterface;
+use unclead\multipleinput\renderers\TableRenderer;
+use unclead\multipleinput\renderers\RendererInterface;
 
 
 /**
@@ -98,12 +98,12 @@ class MultipleInput extends InputWidget
 
     /**
      * @var string the name of column class. You can specify your own class to extend base functionality.
-     * Defaults to `yii\multipleinput\MultipleInputColumn`
+     * Defaults to `unclead\multipleinput\MultipleInputColumn`
      */
     public $columnClass;
 
     /**
-     * @var string the name of renderer class. Defaults to `yii\multipleinput\renderers\TableRenderer`.
+     * @var string the name of renderer class. Defaults to `unclead\multipleinput\renderers\TableRenderer`.
      * @since 1.4
      */
     public $rendererClass;