+++ tests/unit/TestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace yii\multipleinput\tests\unit;
+namespace unclead\multipleinput\tests\unit;
 
 
 abstract class TestCase extends \PHPUnit_Framework_TestCase