+++ tests/unit/data/TestModel.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace yii\multipleinput\tests\unit\data;
+namespace unclead\multipleinput\tests\unit\data;
 
 use yii\base\Model;
 
 /**
  * Class TestModel
- * @package yii\multipleinput\tests\unit\data
+ * @package unclead\multipleinput\tests\unit\data
  */
 class TestModel extends Model
 {